Use ollama instead of Open AI's APIs
Remove support for Open AI's paid APIs in favor of ollama.

The rationale is that for self-hosting, ollama and modern models are
competent enough to work, even on low-end devices.

If the use case or need for using Open AI arises, it can easily be added
again as the code is written in an extensible way.