Start work on moving to using yt-dlp
We lost the fight with YouTube and getting hold of media is now much
more difficult, to the point that we cannot really do it anymore without
relying on yt-dlp and a JS runtime.

This comes at a rather big performance impact as well as an image size
growth. The positives are that we can now support essentially anything
yt-dlp supports. To keep things performant, we still don't use ffmpeg,
though - so WebM and OPUS is the way to go.